fixes #16
Make-HashTableFromXML function now has an extra switch -rigorous which
by default makes the merging of hashtables much faster (~5sec per
hashtable). What made the merging so slow was the Test-Path for each
hash referece. This is now checked during backup in a script block - and
will raise a warning if file reference doesn't exist.

Merge-process could be made faster, but this would probably require the
use of compiled C-code and Add-Type?